Initial commit of `store` and `worker` modules The `store` module is a basic abstraction for the database, which as of right now is still just a hash map that's owned by a single thread.
`worker` is a proof of concept for a worker that handles interaction with the data store, and communicates with the server's event loop using oneshot channels. I will most likely expand on this later on, to make it so pools of workers can be run simultaneously - this is just laying the groundwork while I learn about some of these concepts, and get my footing with Rust.